Trim to make cluster name shorter (#99)
Instead of initializing the `CLUSTER_NAME` environment variable in the
`.devcontainer/devcontainer.json` file with the value of
`${localEnv:CODESPACE_NAME}`, this pull request uses the
`.devcontainer/postStartCommand.sh` script to set the `CLUSTER_NAME`
environment variable to the value of `${CODESPACE_NAME%-*}` via
`~/.bashrc`. This change shortens the cluster name by removing the
suffix after and including the last hyphen in the `CODESPACE_NAME`
environment variable.

For example, my `CODESPACE_NAME` is `friendly-couscous-p74x5x57qp2xgg`,
and the `CLUSTER_NAME` will be set to `friendly-couscous` after this
change.

The change also makes the cluster name match the codespace name as
displayed by VS Code


![image](https://github.com/user-attachments/assets/8640b7ea-6d19-4116-8ea3-c11c332b036e)

This makes the cluster name significantly shorter and probably means
that we don't need to trim the Event Hub names any more.