Auto-run artifact pushing if changes detected (#166)
This pull request updates the GitHub Actions workflows to trigger more
precisely on changes relevant to each workflow instead of only manually.

Workflow trigger enhancements:

*
[`.github/workflows/ghcr_graph.yml`](diffhunk://#diff-7fb5501212b99566ed38d9dfe1be93c2e1da9a556a2e9534cad4e203dd7fdd17R7-R16):
Added filters so the workflow only runs on pushes and pull requests to
the `main` branch when files in `samples/wasm/*.yaml` or the workflow
file itself are changed.
*
[`.github/workflows/ghcr_module.yml`](diffhunk://#diff-b5bf96bf48b8c908fb53f124f83ebc1e90c107e8cc9756edb638e4907392cd6aR7-R18):
Added filters so the workflow only runs on pushes and pull requests to
the `main` branch when files in `samples/wasm/operators/**`,
`samples/wasm/Makefile`, or the workflow file itself are changed.